[frontend-5] shared CompanySearch and AreaSearch, added AreaInit
