+++ frontend/src/redux/search/searchReducer.js
@@ -1,13 +1,11 @@
 import {createReducer} from 'redux-create-reducer';
 
-import {FETCH_COMPANY, CHOOSE_COMPANY, FETCH_AREA, CHOOSE_AREA} from './searchActions';
+import {FETCH_COMPANY, CHOOSE_COMPANY} from './searchActions';
 
 
 export const initialState = {
     companyId: undefined,
     companies: undefined,
-    areaId: undefined,
-    areas: undefined
 };
 
 export const searchReducer = createReducer(initialState, {
@@ -23,17 +21,5 @@ export const searchReducer = createReducer(initialState, {
             ...state,
             companyId: action.companyId,
         };
-    },
-    [FETCH_AREA](state, action) {
-        return {
-            ...state,
-            areas: action.filteredAreas,
-        }
-    },
-    [CHOOSE_AREA](state, action) {
-        return {
-            ...state,
-            areaId: action.areaId,
-        }
     },
 });
